Imports rewrite in Ansible
Changed: Depreciated the monitorall flag, it was redundant and needs to
die in a fire.

All of the operations with remote_syslog gem are now controlled with a
master yaml file that lives in /etc/log_files.yml

SyslogNG and RSyslog support has been abstracted into ansible plays
instead of wonky conditionals

Adds unit tests as part of ansible charm scaffolding, however these are
not implemented. left on the TODO list.